Crate re_entity_db

This is how we store and index logging data.

Feature flags

serde — Enable (de)serialization using serde.

Re-exports

pub use self::entity_db::EntityDb;

pub use self::entity_db::DEFAULT_GC_TIME_BUDGET;

pub use self::entity_tree::EntityTree;

pub use re_log_types::EntityPath;

pub use re_log_types::EntityPathPart;

pub use re_log_types::TimeInt;

pub use re_log_types::Timeline;

Modules

entity_db

entity_tree

external

instance_path 🔒

store_bundle 🔒

time_histogram_per_timeline 🔒

times_per_timeline 🔒

versioned_instance_path 🔒

Structs

InstancePath

The path to either a specific instance of an entity, or the whole entity.

InstancePathHash

Hashes of the components of an InstancePath.

StoreBundle

Stores many EntityDbs of recordings and blueprints.

TimeHistogramPerTimeline

Number of messages per time per timeline.

TimesPerTimeline

A [ChunkStoreSubscriber] that keeps track of all unique timestamps on each Timeline.

VersionedInstancePath

A versioned path (i.e. pinned to a specific [RowId]) to either a specific instance of an entity, or the whole entity.

VersionedInstancePathHash

Hashes of the components of a VersionedInstancePath.

Enums

Error

The errors that can occur when misusing the chunk store.

StoreLoadError

TimeQuery

A query in time.

Type Aliases

Result

TimeCounts

TimeHistogram

Number of messages per time.