Module re_entity_db::entity_tree

Structs

ClearCascade

Cascaded clears that need be to applied to the DataStore as a result of modifying the EntityTree.

CompactedStoreEvents

Maintains an optimized representation of a batch of StoreEvents specifically designed to accelerate garbage collection of EntityTrees.

EntityInfo

Information about this specific entity (excluding children).

EntityTree

A recursive, manually updated re_data_store::StoreSubscriber that maintains the entity hierarchy.

SubtreeInfo

Info about stuff at a given EntityPath, including all of its children, recursively.