Crate re_entity_db

This is how we store and index logging data.

Feature flags

serde — Enable (de)serialization using serde.

Re-exports

pub use self::entity_db::EntityDb;

pub use self::entity_tree::EntityTree;

pub use blueprint::components::EntityPropertiesComponent;

pub use self::entity_properties::*;

Modules

blueprint

editable_auto_value 🔒

entity_db

entity_properties

entity_tree

external

instance_path 🔒

store_bundle 🔒

time_histogram_per_timeline 🔒

times_per_timeline 🔒

versioned_instance_path 🔒

Structs

EntityPath

The unique identifier of an entity, e.g. camera/3/points

EntityPathPart

The different parts that make up an EntityPath.

InstancePath

The path to either a specific instance of an entity, or the whole entity.

InstancePathHash

Hashes of the components of an InstancePath.

StoreBundle

Stores many EntityDbs of recordings and blueprints.

TimeHistogramPerTimeline

Number of messages per time per timeline.

TimeInt

A 64-bit number describing either nanoseconds, sequence numbers or fully static data.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

TimesPerTimeline

A StoreSubscriber that keeps track of all unique timestamps on each Timeline.

VersionedInstancePath

A versioned path (i.e. pinned to a specific RowId) to either a specific instance of an entity, or the whole entity.

VersionedInstancePathHash

Hashes of the components of a VersionedInstancePath.

Enums

EditableAutoValue

A value that is either determined automatically by some heuristic, or specified by the user.

Error

The errors that can occur when misusing the data store.

StoreLoadError

TimeQuery

A query in time.

Type Aliases

Result

TimeCounts

TimeHistogram

Number of messages per time.